feature: refactor email subsystem
