+++ AIMS/src/main/java/com/itep/hust/aimsgroup/model/account/Account.java
@@ -68,8 +68,6 @@ public class Account {
         this.status = status;
     }
 
-
-
     public Account() {
 
     }
