+++ AIMS/src/main/java/com/itep/hust/aimsgroup/model/account/Role.java
@@ -3,10 +3,7 @@ package com.itep.hust.aimsgroup.model.account;
 import java.util.Objects;
 
 public class Role {
-
     private String roleName;
-
-
 
     public String getRoleName() {
         return roleName;
