+++ AIMS/src/main/java/com/itep/hust/aimsgroup/subsystem/email/EmailSubsystem.java
@@ -1,6 +1,6 @@
 package com.itep.hust.aimsgroup.subsystem.email;
 
-public abstract class EmailService {
+public abstract class EmailSubsystem {
     public void sendMail(String recipient, String subject, String content) {
         new Thread(() -> {
             send(recipient, subject, content);
