+++ AIMS/src/main/java/com/itep/hust/aimsgroup/view/invoice/InvoiceViewHandler.java
@@ -4,7 +4,7 @@ import com.itep.hust.aimsgroup.controller.placeorder.PlaceOrderController;
 import com.itep.hust.aimsgroup.model.cart.Cart;
 import com.itep.hust.aimsgroup.model.invoice.Invoice;
 import com.itep.hust.aimsgroup.model.media.Media;
-import com.itep.hust.aimsgroup.subsystem.email.javax.JavaxEmailService;
+import com.itep.hust.aimsgroup.subsystem.email.javax.JavaxEmailSubsystem;
 import com.itep.hust.aimsgroup.util.Screen;
 import com.itep.hust.aimsgroup.view.deliveryinfo.DeliveryInfoViewHandler;
 import javafx.beans.property.ReadOnlyObjectWrapper;
@@ -60,7 +60,7 @@ public class InvoiceViewHandler {
 
     private void initializePaymentButton() {
         paymentButton.setOnMouseClicked(e -> {
-            placeOrderController.redirectToPayment(invoice, new JavaxEmailService());
+            placeOrderController.redirectToPayment(invoice, new JavaxEmailSubsystem());
         });
     }
